zXczxczxc

+++ database/migrations/2023_07_12_082418_create_products_table.php
@@ -20,8 +20,8 @@ class CreateProductsTable extends Migration
             $table->integer('id_user_update');
             $table->string('title');
             $table->string('description');
-            $table->integer('amount');
-            $table->integer('price');
+            $table->bigInteger('amount');
+            $table->bigInteger('price');
             $table->timestamps();
         });
     }
