+++ database/migrations/2023_07_19_095751_add_blance_to_users_table.php
@@ -14,7 +14,7 @@ class AddBlanceToUsersTable extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->integer('balance')->default(0);
+            $table->bigInteger('balance')->default(0);
         });
     }
