+++ database/migrations/2023_07_25_024955_create_carts_table.php
@@ -31,7 +31,7 @@ class CreateCartsTable extends Migration
             $table->bigInteger('amount');
             $table->bigInteger('price_cart')->nullable();
             $table->bigInteger('order_number')->nullable();
-            $table->integer('id_product_choese')->nullable();
+            $table->bigInteger('id_product_choese')->nullable();
             $table->timestamps();
         });
     }
